my third commit:hhh who cares about that comment btw

+++ routes/web.php
@@ -24,12 +24,12 @@ Route::get('/convertibles', [ConvertiblesController::class, 'index']);
 Route::get('/order', function () {
     return view('order');
 });
-Route::get('/cars/create', [CarController::class, 'create']);
-Route::post('/cars', [CarController::class, 'store']);
+Route::get('/cars/create', [CarController::class, 'create'])->middleware('auth', 'admin');
+Route::post('/cars', [CarController::class, 'store'])->middleware('auth', 'admin');
 Route::get('/car/{id}', [CarController::class, 'show'])->name('car.show');
-Route::get('/cars/{id}/edit', [CarController::class, 'edit'])->name('cars.edit');
-Route::put('/cars/{id}', [CarController::class, 'update'])->name('cars.update');
-Route::delete('/cars/{id}', [CarController::class, 'destroy'])->name('cars.destroy');
+Route::get('/cars/{id}/edit', [CarController::class, 'edit'])->name('cars.edit')->middleware('auth', 'admin');
+Route::put('/cars/{id}', [CarController::class, 'update'])->name('cars.update')->middleware('auth', 'admin');
+Route::delete('/cars/{id}', [CarController::class, 'destroy'])->name('cars.destroy')->middleware('auth', 'admin');
 
 Route::get('/login', [LoginController::class, 'showLoginForm'])->name('login');
 Route::post('/login', [LoginController::class, 'login']);
